% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{rectangle}
\alias{rectangle}
\title{rectangle}
\usage{
rectangle(x, y, w, h, ...)
}
\arguments{
\item{x}{Numeric lower left corner x coordinate}

\item{y}{Numeric lower left corner y coordinate}

\item{w}{Numeric width of the rectiangle}

\item{h}{Numeric height of the rectiangle}

\item{...}{Arguments to be fed to \code{polygon}}
}
\description{
Add a rectangle to a plot.
}
\examples{
plot.new()
plot.window(xlim = c(-1, 1), ylim = c(-1, 1))
circle(-0.5, -0.5, 1, 1)
}

