% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{setAlpha}
\alias{setAlpha}
\title{setAlpha}
\usage{
setAlpha(col, alpha)
}
\arguments{
\item{col}{A character specifying a standard R colour}

\item{alpha}{A numeric in [0, 1] specifying the alpha level}
}
\value{
A character vector with elements of 7 or 9 characters, "#" followed by the red, blue, green and optionally alpha values in hexadecimal (after rescaling to 0 ... 255).
}
\description{
Set the alpha level of a standard R colour.
}
\examples{

plot(
  rnorm(10000),
  rnorm(10000),
  pch = 16,
  col = setAlpha("slateblue", 0.2)
)
}

